feat: separate PBH transaction utils into test crate
