+++ world-chain-builder/crates/world/node/tests/mod.rs
@@ -25,14 +25,13 @@ use reth_optimism_node::{OpNetworkPrimitives, OpPayloadBuilderAttributes};
 use reth_primitives_traits::SignedTransaction;
 use reth_provider::providers::BlockchainProvider;
 use revm_primitives::{Address, FixedBytes, B256, U256};
+use world_chain_builder_test_utils::utils::{signer, tree_root};
+use world_chain_builder_test_utils::{DEV_WORLD_ID, PBH_DEV_ENTRYPOINT, PBH_DEV_SIGNATURE_AGGREGATOR};
 use std::collections::BTreeMap;
 use std::ops::Range;
 use std::sync::Arc;
 use world_chain_builder_pool::ordering::WorldChainOrdering;
 use world_chain_builder_pool::root::LATEST_ROOT_SLOT;
-use world_chain_builder_pool::test_utils::{
-    signer, tree_root, PBH_TEST_ENTRYPOINT, PBH_TEST_SIGNATURE_AGGREGATOR, TEST_WORLD_ID,
-};
 use world_chain_builder_pool::tx::WorldChainPooledTransaction;
 use world_chain_builder_pool::validator::WorldChainTransactionValidator;
 use world_chain_builder_rpc::{EthApiExtServer, WorldChainEthApiExt};
@@ -106,9 +105,9 @@ impl WorldChainBuilderTestContext {
             builder_args: WorldChainBuilderArgs {
                 num_pbh_txs: 30,
                 verified_blockspace_capacity: 70,
-                pbh_entrypoint: PBH_TEST_ENTRYPOINT,
-                signature_aggregator: PBH_TEST_SIGNATURE_AGGREGATOR,
-                world_id: TEST_WORLD_ID,
+                pbh_entrypoint: PBH_DEV_ENTRYPOINT,
+                signature_aggregator: PBH_DEV_SIGNATURE_AGGREGATOR,
+                world_id: DEV_WORLD_ID,
             },
             ..Default::default()
         };
@@ -293,7 +292,7 @@ fn get_chain_spec() -> OpChainSpec {
 
     OpChainSpecBuilder::base_mainnet()
         .genesis(genesis.extend_accounts(vec![(
-            TEST_WORLD_ID,
+            DEV_WORLD_ID,
             GenesisAccount::default().with_storage(Some(BTreeMap::from_iter(vec![(
                 LATEST_ROOT_SLOT.into(),
                 tree_root().into(),
