+++ world-chain-builder/crates/world/test-utils/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "world-chain-builder-test-utils"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[dependencies]
+world-chain-builder-pbh.workspace = true
+
+reth-optimism-node.workspace = true
+reth-primitives.workspace = true
+reth-optimism-primitives.workspace = true
+
+alloy-consensus.workspace = true
+alloy-primitives.workspace = true
+alloy-rpc-types.workspace = true
+alloy-eips.workspace = true
+alloy-sol-types.workspace = true
+alloy-rlp.workspace = true
+alloy-signer-local.workspace = true
+alloy-signer.workspace = true
+alloy-network.workspace = true
+op-alloy-consensus.workspace = true
+
+# 3rd party
+semaphore.workspace = true
+bon.workspace = true
+serde.workspace = true
+
+[dev-dependencies]
+test-case.workspace = true
+
+[lints]
+workspace = true
